requirements.txt
fastapi
uvicorn
numpy
python-docx
python-multipart
pymupdf4llm